status_notification added
status_notification is added and should work but needs to be tested with the back-end for confirmation on that. Will be tested as soon as back-end is ready. Timestamp might need to be changed depending on how back-end want this formated. An unused variable named formated_timestamp exists to make this easily!