Seperate the statemachine from the websocket
The websocket is now mostly seprated from the statemachine which helped make the statemachine alot more responsive. Some bugs were fixed, and serialNumber is updated.

voltage commands where changed after a meeting with BMS

chargerID was sent as a "connectorID" so it was changed in websocket_communcation

The new updates should not affect the websocket communication and the code was tested on the hardware